Added basic nx setup with example nextjs and ui
