+++ apps/example-nextjs-app/.eslintrc.json
