+++ apps/example-nextjs-app/public/favicon.ico
