+++ apps/example-nextjs-app/src/pages/api/hello.ts
