+++ apps/example-nextjs-app/src/pages/dev.tsx
@@ -1,0 +1,9 @@
+import { Button } from 'example-ui-lib';
+
+export default function Dev() {
+  return (
+    <div>
+      <Button onClick={() => console.log('clicked')}>Click me</Button>
+    </div>
+  );
+}
