+++ libs/example-ui-lib/Button.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+// packages/shared-ui/Button.tsx
+export function Button(props: any) {
+    return <button onClick={() => props.onClick()}>{props.children}</button>;
+}
+
+export default Button;
