+++ libs/example-ui-lib/index.tsx
@@ -1,0 +1,1 @@
+export * from './Button';
