+++ libs/example-ui-lib/package.json
@@ -1,15 +1,15 @@
 {
-  "name": "shared-ui",
-  "version": "1.0.0",
+  "name": "example-ui-lib",
   "description": "",
-  "main": "index.js",
+  "private": true,
+  "version": "0.0.0",
+  "main": "dist/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "rm -rf dist && tsc"
   },
   "keywords": [],
   "author": "Stamati Morellas",
   "license": "ISC",
-  "private": true,
   "dependencies": {
     "react": "18.2.0"
   },
