+++ nx.json
@@ -1,0 +1,10 @@
+{
+    "tasksRunnerOptions": {
+        "default": {
+            "runner": "nx/tasks-runners/default",
+            "options": {
+                "cacheableOperations": ["build", "test", "lint"]
+            }
+        }
+    }
+}
